requirements.txt
click==7.1.2
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.20.2
Pillow==8.2.0
pkg-resources==0.0.0
